requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
openai>=1.68.0
numpy==2.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
pypdf==5.4.0
python-docx==1.1.2
